Move dispense listener into TNT tracking
