+++ src/main/java/net/countercraft/movecraft/combat/MovecraftCombat.java
@@ -19,7 +19,6 @@ import net.countercraft.movecraft.combat.features.tracking.DamageTracking;
 import net.countercraft.movecraft.combat.features.tracking.FireballTracking;
 import net.countercraft.movecraft.combat.features.tracking.TNTTracking;
 import net.countercraft.movecraft.combat.listener.CraftCollisionExplosionListener;
-import net.countercraft.movecraft.combat.listener.DispenseListener;
 import net.countercraft.movecraft.combat.listener.ExplosionListener;
 import net.countercraft.movecraft.combat.localisation.I18nSupport;
 import org.bukkit.plugin.java.JavaPlugin;
@@ -82,7 +81,6 @@ public final class MovecraftCombat extends JavaPlugin {
 
         // Register event translation listeners
         getServer().getPluginManager().registerEvents(new CraftCollisionExplosionListener(), this);
-        getServer().getPluginManager().registerEvents(new DispenseListener(), this);
         getServer().getPluginManager().registerEvents(new ExplosionListener(), this);
